+++ plugins/watcher/package.js
@@ -1,6 +1,6 @@
 /* eslint-disable prefer-arrow-callback */
 Package.describe({
-    name: 'communitypackages:meteor-desktop-watcher',
+    name: 'skadmin:meteor-desktop-watcher',
     version: '3.0.1',
     summary: 'Watches .desktop dir and triggers rebuilds on file change.',
     git: 'https://github.com/sharekey/meteor-desktop',
@@ -16,7 +16,7 @@ Package.onUse(function onUse(api) {
     api.versionsFrom('METEOR@1.3.3');
     api.use('ecmascript');
     api.use([
-        'communitypackages:meteor-desktop-bundler@3.0.0',
+        'skadmin:meteor-desktop-bundler@3.0.0',
     ], ['server'], {
         weak: true
     });